Keymap: Add duration to get connection state after BMP event received
In the current implementation, if adv command is triggerd without id argument,
BLE_CONNECTED event is generated before the connection status is updated
in the firmware. If this issue occurred, wrong device info is displayed on OLED.
Adding duration about 1.5 sec to wait status update in the firmware to
avoid this issue.