rgblight: Fix wrong lottery logic in twinkle rgblight
rand() returns an int type value which is usually a very
large value compared with the threshold in the current lottery
with uint8_t or uint16_t. Because of this, most of the lottery
fails and no twinkle is observed on the keyboard.

Cherry-pick from original qmk repository: 571a9794af0249ef23a0af2ed2ee64191299faa7