Matrix: Forming frames and flush data in uart communication
UART SDK in BMP internally buffers received data and passes it
to the user application every 128 bytes. In order to ensure the
timly communication, send additional 128 byte after the data
transmission to flush the buffer.
On the other hand, data framing is also needed in the receiver
to find the start point of the data structure.